requirements.txt
httpx
openai
python-dotenv
beautifulsoup4
lxml